+++ example/notifier.dart
@@ -1,0 +1,8 @@
+import 'dart:io';
+
+import 'package:colorful_cmd/component.dart';
+
+void main(List<String> args) {
+  var notifierProxy = NotifierProxy(mac: [TerminalNotifier(), AppleScriptNotifier()], linux: [NotifySendNotifier()], win: [NotifuNotifier()]);
+  notifierProxy.send('This is notify!', title: 'notify title', subtitle: 'notify subtitle', groupID: 'colorful_cmd', openURL: 'http://github.com/AlanAlbert');
+}