+++ lib/src/interface/i_cmd.dart
@@ -117,7 +117,10 @@ abstract class ICmd<T> extends Command<T> {
   }
 
   /// print info
-  void info(Object text, [Color color = Color.CYAN]) => printInfo(text, color);
+  void info(Object text, [Color color]) {
+    color ??= Colors.CYAN;
+    printInfo(text, color);
+  }
 
   void debug(Object text) => printDebug(text);
 
@@ -135,7 +138,7 @@ abstract class ICmd<T> extends Command<T> {
 
   String _getTitle() {
     if (runner.showTitle) {
-      return TextPen()
+      return ColorText()
           .setColor(runner.titleColor)
           .text('${formatChars(runner.executableName)}\n')
           .normal()